Typescript migration, refs #113
